# Fix multiple issues: CodeRabbit rules, data export flow, test runner, email logging, and test utilities