… .augment/rules/Coderabbit-auto-fixes.md
… .augment/rules/augment-rules.md
… scripts/test/runner.js
… src/services/emailService.js
… src/templates/emails/data_export.hbs
… src/utils/logger.js
… src/workers/BaseWorker.js
… tests/helpers/testUtils.js
… tests/unit/frontend/Settings-Profile-logic.test.js
… tests/unit/routes/user-profile-simple.test.js
… tests/unit/utils/safeUtils.test.js
… tests/unit/utils/testUtils-planLimits.test.js
… tests/unit/workers/BaseWorker.test.js