@Eibon7 2025-08-29
## ✅ Tarea Completada

Todos los issues solicitados han sido resueltos exitosamente:

### 🔧 Cambios Implementados:
1. **CodeRabbit auto-fixes rule** - Actualizada con cierre de issues opt-in y requisitos de variables de entorno
2. **Settings.jsx data export flow** - Unificado al flujo de modal basado en email
3. **Test runner CLI command** - Nuevo comando 'check' con soporte JSON
4. **User.js rate limit docstring** - Actualizado de 3 a 5 intentos por hora
5. **Email logging privacy** - Implementado SafeUtils.maskEmail para protección PII
6. **Duplicate emailService import** - Eliminado import duplicado
7. **TestUtils.js plan resolution** - Corregida resolución de planes y entitlements
8. **Supabase mock** - Añadido createUserClient mock faltante

### 🧪 Tests Verificados:
- ✅ User Profile Simple Tests: 9/9 pasando
- ✅ SafeUtils Tests: 15/15 pasando (incluyendo nueva función maskEmail)
- ✅ TestUtils Plan Limits Tests: Todos pasando
- ✅ Nuevo comando check: Funcionando con salida JSON y legible

### 📋 Próximos Pasos:
Se ha creado una nueva tarea para abordar los 80 test suites fallidos pre-existentes (no relacionados con estos cambios) para mejorar la calidad general del código.

Todos los cambios mantienen compatibilidad hacia atrás y han sido probados exitosamente. 🚀