@Eibon7 2025-08-29
## ✅ **Job Arreglado Exitosamente**

**Estado del workflow "Runner JSON Demo":**
- ✅ **Antes**: Fallaba con error de parsing JSON (no podía leer `lintExitCode`/`typecheck.exitCode`)
- ✅ **Ahora**: Ejecuta correctamente y parsea el JSON del nuevo comando `check`
- ✅ **Comportamiento actual**: Falla con exit code 1 porque los tests tienen fallos (comportamiento esperado)

### 🔍 **Verificación:**

**Comando funcionando localmente:**
```bash
# JSON output
node scripts/test/runner.js check --json
# → {"command":"npm run test","exitCode":1,"stdout":"...","stderr":"..."}

# Human-readable output  
node scripts/test/runner.js check
# → 🔍 Ran: npm run test
# → Exit code: 1
# → 📤 Output: [test results]
```

**Workflow ahora:**
- ✅ Parsea correctamente `exitCode` del JSON
- ✅ Muestra el comando ejecutado
- ✅ Maneja errores apropiadamente
- ✅ Falla con exit code 1 cuando los tests fallan (comportamiento correcto)

### 🎆 **Resultado:**
El job ya no tiene errores de parsing y funciona como se esperaba. El workflow falla porque hay tests con problemas pre-existentes, pero eso es el comportamiento correcto - el CI debe fallar cuando hay tests fallidos.

**¡Fix completado exitosamente!** 🚀